✨ feat(config): Add release rules for minor, patch and major releases based on commit type
This commit introduces new release rules to the configuration of semantic-release. The release rules are based on the type of the commit messages. Commits with the prefix "feat" will trigger a minor release, commits with the prefix "fix" or "perf" will trigger a patch release, and commits with the prefix "style", "refactor", or "build" will also trigger a patch release. In addition, there are specific rules for documentation changes and breaking changes. Finally, two special keywords "[force release]" and "[skip release]" have been added to control the release process manually.